Make the commands self-registering on application startup
So that CommandFactory::registerCommands() dpesn't have to have a big
list of all the known commands (and be updated when a new one is added),
and doesn't need to create an instance of every command.

This looks like a big change, but the changes to the individual commands are
minimal;  the DEFINE_COMMAND macro does all of the work.