Handle CommandShell the same way as the other commands
So that it doesn't need to be a special case in main() and elsewhere.
It still needs to be handled specially in some places, but minimally.

This means that executing 'akonadiclient' with no arguments gives a
meaningful error message instead of entering the shell (with no
prompt, so giving the impression that the command is hanging).
Use 'akonadiclient shell' to enter the command shell.