Update command 'expand' to also be able to modify the group
It can now operate in four modes: expand (as before, the default),
add, delete or clean (invalid references from the group).  There
are new command line options to select which of these is to be done.
Items to be added or deleted can be specified by either item
reference or email address (with optional full name).

Renamed the command to 'group' to better reflect its function.