Add 'dump' command
This recursively dumps the items within a collection to a specified
destination in their native format - this means RFC822-format files for
emails, VCF files for contacts etc.  The folder structure is maintained;
for emails there is the option to dump either in Maildir format (with
hidden folders, etc) or in plain directories.  Each item is written to a
file identified by its Akonadi ID.

In theory, the resulting output could be used by the 'add' command or by
a PIM application's import facility to restore the original data, even
into a different instance of Akonadi.  In order to achieve this, Akonadi
tags in contact entries are converted to their textual form.