Command 'dump': Add option for categories as internal Akonadi URLs
What is wanted in the output depends on the use case.  For importing the
data into a foreign PIM application then the textual categories are more
useful.  For importing into another instance of Akonadi the internal
URLs allow the category information.  In both cases the Akonadi UID is
hidden, it is not useful even for an Akonadi import.

Also detect MIME type "text/vcard" (used in Akonadi 5) as well as the
original "text/directory".