Big rewrite: port KCmdLineArgs etc -> QCommandLineParser
Pass the QCommndLineParser to AbstractCommand::setupCommandOptions() to
set up the command options, and to AbstractCommand::initCommand() to
get values and options from the parsed arguments.  Make the former
function pure virtual because every command subclass needs it, and
automatically do the base class work first.

Simplify the looping logic and signals/slots in CommandShell by using a
local event loop - not really a problem using that here, because there
is no asynchronous GUI or DBus stuff going on.

Use i18n() instead of ki18n() for all GUI strings, this can be done now
because the QCoreApplication object is created right at the start.