Add 'folders' command to save and check folder ID assignments
Intended for use when the Akonadi database has had to be regenerated,
which means that folder IDs saved in configuration files may be out
of date.  When the database and configuration is stable, the current
folder assignments can be saved by running the command in 'backup' mode.

If the database has to be regenerated later on then the new current
folder assignments can be checked against the backup by running the
command in the 'check' mode.  This updates configuration files but
does not actually restore the active configuration in use; the 'restore'
mode will do that later.