attributes: Add 'backup', 'check' and 'restore' modes
This will do that, in the same way as the 'folders' command, for
collection attributes which are stored in the Akonadi database
and therefore may be lost if the database has to be regenerated.
For example, the folder icons and expiry settings in KMail are
stored in this way.

This so far is the 'backup' operation - saving the current attributes,
for all collections that have them, to a file.