tags: Reinstate the option to allow a tag ID to be assigned
This option is not intended to be used as a matter of course because
the tag ID, in the same way as item and collection IDs, is assigned
as a unique incrementing index by the Akonadi server.  The only way
to assign a particular ID to a newly created tag is to keep on
creating new tags until the ID reaches the requested one - which is
only possible if the ID is not and never has been already assigned.
If the ID has already been assigned then the only way to restart the
sequence is to start again with a clean Akonadi database.  So this
option is only intended for use in situations where the tag ID must
correspond with some external data source that cannot be changed.

The same strategy will be used, with the assumption that the database
is empty, for backing up and restoring the tags.