Add support for JSON output to certian commands
This is useful if the output from the tool is to be passed down
to other CLI tools or wrapped and called e.g. from Python, where
parsing a JSON is much easier and reliable than decoding the custom
text output format. It also allows us to include more information
about each entity in the output.